requirements.txt
Django==3.1.2
djangorestframework==3.11.2
drf-yasg==1.17.1
Pillow==8.0.1
black==20.8b1
